Rename /views to /pages
